requirements.txt
fastapi>=0.103.1
uvicorn[standard]>=0.23.2
pydantic>=2.3.0
sqlalchemy>=2.0.20
alembic>=1.12.0
asyncpg>=0.28.0
python-dotenv>=1.0.1
redis>=5.0.0
tenacity>=8.2.3
pybreaker>=1.0.1
prometheus-fastapi-instrumentator>=0.4.1
playwright>=1.37.0
pytest>=7.4.2
httpx>=0.24.1
websockets>=11.0.3
